factorizar arquitectura intento 1
