+++ src/com/grupo11/dulcehogar/s3/acceso_datos/Conexion.java
@@ -1,4 +1,4 @@
-package bd_dulcehogar;
+package com.grupo11.dulcehogar.s3.acceso_datos;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -15,9 +15,9 @@ public class Conexion {
             Class.forName("com.mysql.cj.jdbc.Driver");
 
             // Datos de conexión
-            String url = "jdbc:mysql://localhost:3306/bd_dulcehogar"; // Asegúrate de que esta URL sea correcta
+            String url = "jdbc:mysql://localhost:3306/dulcehogar"; // Asegúrate de que esta URL sea correcta
             String user = "root";
-            String pass = "Mkg299..";
+            String pass = "oTfuhT_em3dnwyC";
 
             // Establecer la conexión
             conexion = DriverManager.getConnection(url, user, pass);
