+++ src/com/grupo11/dulcehogar/s3/negocio/NegCuentaSocio.java
@@ -1,0 +1,13 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package com.grupo11.dulcehogar.s3.negocio;
+
+/**
+ *
+ * @author Lucas
+ */
+public class NegCuentaSocio {
+    
+}
