+++ src/com/grupo11/dulcehogar/s3/negocio/Socio.java
@@ -1,4 +1,4 @@
-package dulcehogar;
+package com.grupo11.dulcehogar.s3.negocio;
 
 public class Socio {
