+++ src/com/grupo11/dulcehogar/s3/negocio/Validar.java
@@ -1,4 +1,4 @@
-package dulcehogar;
+package com.grupo11.dulcehogar.s3.negocio;
 
 public class Validar {
