+++ src/com/grupo11/dulcehogar/s3/negocio/VentanaPrincipal.java
@@ -1,4 +1,4 @@
-package dulcehogar;
+package com.grupo11.dulcehogar.s3.negocio;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
